Add copyright notice to Readme-main.md
## 📜 Copyright Notice

© 2025 CMD-SYS Project Contributors. All rights reserved.

This software and its associated documentation files are protected under international copyright laws.  
Unauthorized copying, distribution, modification, or use of this project, in whole or in part, without prior written permission from the authors is strictly prohibited.

You are granted permission to use this project under the terms of the **MIT License**.  
This means you may freely use, modify, and distribute the software, provided that:

- Proper credit is given to the original authors.  
- The copyright notice and license text remain included in all copies or substantial portions of the software.  
- The software is provided “as is,” without warranty of any kind, express or implied.

For full details, please see the [LICENSE](LICENSE) file included in this repository.